+++ src/File.php
@@ -14,6 +14,7 @@ interface File
     public function name(): Name;
     public function content(): Content;
     public function mediaType(): MediaType;
+    public function rename(Name $name): self;
 
     /**
      * This method called on a directory DOES NOTHING